Move theme updates from xaml island template to composition template
