+++ templates/Wpf/Features/ThemeSelection/Services/ThemeSelectorService.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Windows;
-using System.Windows.Media;
 using ControlzEx.Theming;
 using Microsoft.Win32;
 using Param_RootNamespace.Contracts.Services;
@@ -12,8 +11,6 @@ namespace Param_RootNamespace.Services
     {
         private bool IsHighContrastActive
                         => SystemParameters.HighContrast;
-
-        public event EventHandler ThemeChanged;
 
         public ThemeSelectorService()
         {
@@ -60,9 +57,6 @@ namespace Param_RootNamespace.Services
             Enum.TryParse(themeName, out AppTheme theme);
             return theme;
         }
-
-        public SolidColorBrush GetColor(string colorKey)
-            => Application.Current.FindResource(colorKey) as SolidColorBrush;
 
         private void OnUserPreferenceChanging(object sender, UserPreferenceChangingEventArgs e)
         {
