+++ templates/Wpf/Features/XamlIslandUwpApp/Param_ProjectName.XamlIsland/Properties/Param_ProjectName.XamlIsland.rd.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <!--
     This file contains Runtime Directives, specifications about types your application accesses
     through reflection and other dynamic code patterns. Runtime Directives are used to control the
