+++ templates/Wpf/Features/XamlIslandUwpApp/Param_ProjectName.XamlIslandApp/Properties/Default.rd.xml
@@ -1,4 +1,4 @@
-<!--
+﻿<!--
     This file contains Runtime Directives used by .NET Native. The defaults here are suitable for most
     developers. However, you can modify these parameters to modify the behavior of the .NET Native
     optimizer.
