+++ templates/Wpf/_comp/_shared/Feature.XamlIsland.Theming/.template.config/template.json
@@ -1,0 +1,28 @@
+﻿{
+    "$schema": "http://json.schemastore.org/template",
+  "author": "Microsoft",
+    "classifications": [
+      "Universal"
+    ],
+    "name": "wts.Wpf.Feature.Feature.XamlIsland.Theming",
+    "shortName": "wts.Wpf.Feature.XamlIsland.Theming",
+    "identity": "wts.Wpf.Feature.XamlIsland.Theming",
+  "tags": {
+    "language": "C#",
+    "type": "item",
+    "wts.type": "composition",
+    "wts.platform": "Wpf",
+    "wts.version": "1.0.0",
+    "wts.compositionFilter": "identity == wts.Wpf.Page.XamlIsland & $feature == wts.Wpf.Feat.ThemeSelection"
+  },
+    "sourceName": "wts.ItemName",
+    "preferNameDirectory": true,
+    "PrimaryOutputs": [
+    ],
+    "symbols": {
+      "wts.rootNamespace": {
+        "type": "parameter",
+        "replaces": "Param_RootNamespace"
+      }
+    }
+  }
