+++ templates/Wpf/_comp/_shared/Feature.XamlIsland.Theming/Contracts/Services/IThemeSelectorService_postaction.cs
@@ -1,0 +1,17 @@
+﻿//{[{
+using System.Windows.Media;
+//}]}
+
+namespace Param_RootNamespace.Contracts.Services
+{
+    public interface IThemeSelectorService
+    {
+//{[{
+        event EventHandler ThemeChanged;
+
+        SolidColorBrush GetColor(string colorKey);
+
+//}]}
+    }
+}
+
