+++ templates/Wpf/_comp/_shared/Feature.XamlIsland.Theming/Services/ThemeSelectorService_postaction.cs
@@ -1,0 +1,21 @@
+﻿//{[{
+using System.Windows.Media;
+//}]}
+
+namespace Param_RootNamespace.Services
+{
+    public class ThemeSelectorService : IThemeSelectorService
+    {
+//{[{
+        public event EventHandler ThemeChanged;
+
+//}]}
+        public ThemeSelectorService()
+        {
+        }
+//{[{
+        public SolidColorBrush GetColor(string colorKey)
+                    => Application.Current.FindResource(colorKey) as SolidColorBrush;
+//}]}
+    }
+}
