+++ templates/Wpf/_comp/_shared/Project.AddPlatforms/.template.config/template.json
@@ -1,4 +1,4 @@
-{
+﻿{
   "schema": "http://json.schemastore.org/template",
   "author": "Microsoft",
   "classifications": [
